refactor(settings): add user preferences management hook
